powermeter refactor: SML lib: replace double by float
avoid additional conversions and avoid double for the fact that
calculations on type double are implemented in software, whereas
float is handled in hardware on ESP32.

+++ include/PowerMeterSml.h
@@ -27,7 +27,7 @@ private:
 
     typedef struct {
         uint8_t const OBIS[6];
-        void (*decoder)(double&);
+        void (*decoder)(float&);
         float* target;
         char const* name;
     } OBISHandler;
